+++ src/main/frontend/src/app/mapping/reviews-dialog/reviews-dialog.component.ts
@@ -18,7 +18,7 @@
 
 import { Component, Inject, EventEmitter } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { ConceptId, VocabularyId, CodeId } from '../data';
+import { ConceptId, VocabularyId, CodeId } from '../mapping-data';
 import { AllTopics, TopicsInfo, ReviewData, ReviewOperation } from '../review';
 
 export interface ReviewsData {